+++ src/apps/yearn/yearn.definition.ts
@@ -11,6 +11,7 @@ export const YEARN_DEFINITION = appDefinition({
     farm: { id: 'farm', type: GroupType.POSITION, label: 'Governance' },
     v1Vault: { id: 'v1-vault', type: GroupType.TOKEN, label: 'Vaults (V1)' },
     v2Vault: { id: 'v2-vault', type: GroupType.TOKEN, label: 'Vaults' },
+    governance: { id: 'governance', type: GroupType.POSITION, label: 'Governance' },
     yield: { id: 'yield', type: GroupType.TOKEN, label: 'Yield Tokens', isHiddenFromExplore: true }, // We seem to have these twice e.g. the same yUSDT is here and in vaults
   },
   url: 'https://yearn.finance/',